requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
jinja2==3.0.1
aiofiles==0.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.23
pydantic[email]==1.8.2
python-dotenv==0.19.0
requests==2.26.0
beautifulsoup4==4.9.3
aiohttp==3.8.1
pandas==1.4.2 